third_party/WebKit/LayoutTests/external/wpt/css/css-masking/parsing/mask-border-source-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Masking Module Level 1: parsing mask-border-source with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/css-masking-1/#the-mask-border-source">
<meta name="assert" content="mask-border-source supports only the grammar 'none | <image>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("mask-border-source", "auto");
test_invalid_value("mask-border-source", "circle()");
test_invalid_value("mask-border-source", "border-box");
</script>
</body>
</html>
